+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.13.1 - 2023-10-23
+### Enhancements
+- Added new publisher values in preparation for DBEQ.PLUS
+- Added `ToIso8601` for `UnixNanos` for converting to human-readable ISO8601 datetime
+  string
+- Added `kUndefTimestamp` and `kUndefStatQuantity` constants
+- Added flag `kTob` for top-of-book messages
+
 ## 0.13.0 - 2023-09-21
 ### Enhancements
 - Added `pretty_px` option for `BatchSubmitJob`, which formats prices to the correct